V2.8.20 "Breaking change" need to delete command_use.json
