+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "kasuki"
-version = "2.8.19"
+version = "2.8.20"
 edition = "2021"
 authors = ["Valgul <contact@valgul.moe>"]
 description = "A discord bot written in rust that get info from anilist API"
