+++ src/command/autocomplete/anilist_server/delete_activity.rs
@@ -48,7 +48,8 @@ pub async fn autocomplete(ctx: Context, autocomplete_interaction: CommandInterac
     };
 
     let activities = get_data_all_activity_by_server(&guild_id).await.unwrap();
-    let activity: Vec<String> = activities.clone()
+    let activity: Vec<String> = activities
+        .clone()
         .into_iter()
         .map(|activity| format!("{}${}", activity.1, activity.0))
         .collect();
